Move JSON helpers /src/lib/json > /src/lib/utilities
