+++ src/app/containers/StoryPromo/index.stories.jsx
@@ -4,7 +4,7 @@ import { ServiceContextProvider } from '../../contexts/ServiceContext';
 import { RequestContextProvider } from '../../contexts/RequestContext';
 import StoryPromoConatiner from '.';
 import fixture from '../../../../data/prod/pidgin/frontpage';
-import deepGet from '../../lib/json/deepGet';
+import deepGet from '../../lib/utilities/deepGet';
 import AmpDecorator from '../../../../.storybook/helpers/ampDecorator';
 
 const getStoryPromo = platform => (
