+++ src/app/containers/StoryPromo/index.test.jsx
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react';
 import { shouldMatchSnapshot } from '../../../testHelpers';
 import { RequestContextProvider } from '../../contexts/RequestContext';
 import { ServiceContextProvider } from '../../contexts/ServiceContext';
-import deepClone from '../../lib/json/deepClone';
+import deepClone from '../../lib/utilities/deepClone';
 import StoryPromo from '.';
 
 let item;
