+++ src/app/containers/Video/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Helmet from 'react-helmet';
 import Figure from '@bbc/psammead-figure';
-import deepGet from '../../lib/json/deepGet';
+import deepGet from '../../lib/utilities/deepGet';
 import Video from '../../components/Video';
 import Caption from '../Caption';
 import videoMetadata from './videoMetadata';
