+++ src/app/containers/Video/videoMetadata.js
@@ -1,4 +1,4 @@
-import deepGet from '../../lib/json/deepGet';
+import deepGet from '../../lib/utilities/deepGet';
 
 const videoObject = 'VideoObject';
