+++ src/app/lib/utilities/preprocessor/rules/cpstypes.js
@@ -1,4 +1,4 @@
-import deepGet from '../../../json/deepGet';
+import deepGet from '../../deepGet';
 
 const whitelist = ['STY', 'MAP'];
