+++ src/app/lib/utilities/preprocessor/rules/filterEmptyGroupItems.js
@@ -1,4 +1,4 @@
-import deepGet from '../../../json/deepGet';
+import deepGet from '../../deepGet';
 
 const filterEmptyGroupItems = jsonRaw => {
   const json = jsonRaw;
