+++ src/app/lib/utilities/preprocessor/rules/timestamp.test.js
@@ -1,5 +1,5 @@
 import applyTimestampRules from './timestamp';
-import deepClone from '../../../json/deepClone';
+import deepClone from '../../deepClone';
 
 const paragraphBlock = {
   type: 'text',
