+++ src/app/lib/utilities/preprocessor/rules/topstories.js
@@ -1,4 +1,4 @@
-import deepGet from '../../../json/deepGet';
+import deepGet from '../../deepGet';
 
 const squashKeys = [
   'responsive-top-stories',
